some refactoring and added play count
