+++ server/src/ga/GaWorker.ts
@@ -39,12 +39,12 @@ export class GaWorker {
       noFavourateWeight: 0.5,
       musicalFitnessWeight: 1,
       randomInitial: false,
-      populationSize: 7,
+      populationSize: 10,
       selectionSize: 7,
     };
   }
 
-  constructor(config: GaWorkerConfig) {
+  constructor(config: GaWorkerConfig = GaWorker.defaultConfig()) {
     this.config = config;
     this.rules = [
       new LeapRule(),
@@ -107,11 +107,7 @@ export class GaWorker {
     generation.forEach((bot) => (bot.metric = 0));
 
     // Sort bots by fitness, descending, then select most fit bots
-    // TODO: introduce random weighted sorting
-    const selection = _.zip(generation, fitnesses)
-      .sort((a: any[], b: any[]) => b[1] - a[1])
-      .map((botWithFitness) => botWithFitness[0])
-      .slice(0, this.config.selectionSize);
+    const selection = selectRandomWeighted(generation, fitnesses, generation.length);
 
     // Ensure favourited robots are persisted
     startingPopulation.forEach((bot, i) => {
