+++ server/src/utils/Utils.ts
@@ -7,7 +7,7 @@ import { Note } from "../models/Note";
 export const parseBotsFromReq = (req) => {
   return req.body.bots.map(
     (bot) =>
-      new Bot(bot.metric, new Melody(bot.melody.notes.map((note) => Note.fromString(note.note))))
+      new Bot(bot.metric, new Melody(bot.melody.notes.map((note) => Note.fromString(note.note))), bot.playCount)
   );
 };
